oak virtualisation - automatically share pipewire socket with libvirt
Didn't find a systemd based solution that works. Both executing as a
service user with the `users` fs option, and executing as my user with
the `user` fs option failed due to permissions issues in the systemd
user session.

This might have some rough edge if pipewire is restarted, but quick
restarts are fine at least. Seems good enough.