WIP - single GPU passthrough hooks and config
I might revisit this later, but the methods I've seen documented don't
quite seem to be working here. In particular I haven't been able to
unload the nvidia kernel modules. The attempts to remove them complain
that they are in use, but it's not clear why.  None of the framebuffers
in /sys/... are linked, and attempts to unbind them result in a "no such
device" error message.

I'm not sure if there's something different about how nixos handles
framebuffers, or if everything I've read is out of date.