hyprland - replace anyrun launcher with walker
Anyrun has just been a source of a lot of problems. The flake was just
ported to home-manager conflicting with the repo flake, and the
installed anyrun doesn't launch.

I want to replace this with something in vanity eventually, but switching
to walker for now. It's a similar feature set and will tide me over.