feature - add immich wip
This is mostly good to go, but things are behind in the derivation.
Immich migrated to vectorchord in place of pgvecto-rs, and that seems
to be what opened up postgres 17 support. There's already a WIP PR
for this migration: https://github.com/NixOS/nixpkgs/pull/428568

Once this is merged, a few more changes are necessary after user /
service are created:
- Uncomment mediaLocation
- Create zfs dataset for postgres database, migrate initialized data
into that dataset and enable snapshots.