DNS_MIGRATION - temporarily stop rebooting redbud and warden
Migrating to a cloudflare for dns so I can use cf tunnels. External
services should handle this fine, but internally this is used for DNS-01
letsencrypt challenges.

I don't currently persist any acme directories. I also don't really want
to deal with ambiguity around starting and stopping persisting those
directories, so just disabling rebooting for a moment on redbud and
warden. These are currently provisioned new on every reboot, and have
a 90 day lifetime. This is plenty even if something goes wrong.

This could be done with only automated updates, but I want to get it
done faster than that, so I'll be migrating live on oak.