cleanup ntfy tunnel deployment
I thought I needed to terminate TLS at ingress initially. I think I will
still use https for the ingress url if it's on a different server in
this environment, but just keeping it simpler for services on the
same machine as the tunnel.

A little more migration if I move the tunnel to a different server, but
it's a bit cleaner now.