feat(zed) - rework zed notifications to use ntfy wrapper exclusively
I give up on the zed-functions native support for NTFY. It has been such
a pain to debug, especially with how infrequent automatic scrub events
are. Dropping support this and just replacing the old notify cli
with the ntfy cli wrapper I wrote.

Also moving notify to optional. It won't function in the bootstrap iso.

This still needs massaged around a bit more as I figure out exactly
what I want to do with legacyModules, but it's improved from where it
was.